smyrna: make 'viewport_camera.target{x|y}' 'double's
This is part of a series of changes to standardize on `double` arithmetic in
Smyrna. This removes visual noise and improves accuracy. The computational cost
of `double` vs `float` arithmetic is no longer a significant factor in
contemporary environments.

+++ cmd/smyrna/smyrnadefs.h
@@ -129,8 +129,8 @@ enum {SCHEMACOUNT = 4}; // number of colors
     }mouse_action_t;
 
     typedef struct {
-	float targetx;
-	float targety;
+	double targetx;
+	double targety;
 
 	float r;
     } viewport_camera;
